feat(users): add create user action and list users component
Add a new action to create a user and a new component to list all users.
The create user action sends a form data to the server to create a new user.
The list users component displays a list of all users with their name, avatar, position, phone number, and active status.
The list users component also has an "add" button to create a new user if the user has the permission to do so.